#28, Move filter codes into MovieFilter class
